I added Alaska and Hawaii
Why is this change important?: I added my special feature which is to add Alaska and Hawaii. (Alaska has counties, but Hawaii doesn't yet.)
Why did you make the change? It was my special feature to add Alaska and Hawaii and it helped me start to understand how to add the counties.