Colors the states either blue or red or green
This change is important because I was able to get it so that it would color the states. 
I made this change in order to be able to complete the next part of our project. 

I added the coloring code to this. I will need to integrate this into the main class, and will do that once I know that it won't cause any problems to the program.